+++ hpc/metax/samples/hpcc_helpers.h
@@ -1,5 +1,1 @@
-// SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
-#pragma once
-
-void gpu_init(int dev_id);
-const char *gpu_name(int dev_id, char *buf, int buf_len);
+../../nvidia/cuda/samples/cuda_helpers.h